Commit 14: comentado

+++ commands.h
@@ -5,37 +5,44 @@
 #include "function.h"
 #include <string.h>
 
-Cache_ *lru_create(int capacity) // Comando de creación de cache
+Cache_ *lru_create(int capacity) // Comando de creacion de cache
 {
-    if (capacity < 5) {
+    if (capacity < 5)//filtro de capacidad minima
+    {
         printf("La capacidad minima es 5\n");
         capacity = 5;
     }
+
     printf("Creando cache de capacidad %d\n", capacity);    
 
-    Cache_ *cache = (Cache_ *)malloc(sizeof(Cache_));
-    if (cache == NULL) {
+    Cache_ *cache = (Cache_ *)malloc(sizeof(Cache_));//asigna memoria para la estructura Cache_
+    if (!cache)
+    {
         printf("Error al asignar memoria para el cache\n");
         return NULL;
     }
 
-    cache->data = (char **)malloc(capacity * sizeof(char *));
-    if (cache->data == NULL) {
+    cache->data = (char **)malloc(capacity * sizeof(char *));//asigna memoria para los datos del cache
+    if (!cache->data)
+    {
         printf("Error al asignar memoria para los datos del cache\n");
         free(cache);
         return NULL;
     }
 
+    //inicializa los campos de la estructura Cache_
     cache->capacity = capacity;
     cache->size = 0;
     cache->head = NULL;
 
+    //crear carpeta cache
     create_folder();
 
-    
+    //crear archivo data.txt vacio
     FILE *data = fopen("cache/data.txt", "w");
     fclose(data);
 
+    //crear archivo metadata.txt
     update_cache(cache);
     printf("Cache creado con exito\n");
     return cache;
@@ -44,7 +51,7 @@ Cache_ *lru_create(int capacity) // Comando de creación de cache
 Cache_* lru_load_cache() // Carga el cache existente
 {
     Cache_ *cache = (Cache_ *)malloc(sizeof(Cache_));
-    if (cache == NULL) {
+    if (!cache) {
         printf("Error al asignar memoria para el cache\n");
         return NULL;
     }
@@ -63,45 +70,47 @@ Cache_* lru_load_cache() // Carga el cache existente
 
 int lru_add(Cache_ *cache, char *data) // Comando para agregar un dato al cache
 {
-    if(cache == NULL)
+    if(!cache)
     {
         printf("Cache no inicializado. Use 'create <tamano_cache>' para crear uno.\n");
         return -1;
     }
-    if(search_data(cache, data) == 1)
+    if(search_data(cache, data) == 1)//verifica si el dato ya existe en el cache
     {
         printf("El dato '%s' ya existe en el cache. No se agrego.\n", data);
         return -1;
     }
     
-    Node_ *TempNode = (Node_ *)malloc(sizeof(Node_));
+    Node_ *TempNode = (Node_ *)malloc(sizeof(Node_));//crea un nuevo nodo
     if(TempNode == NULL)
     {
         printf("Error al asignar memoria para el nuevo nodo\n");
         return -1;
     }
+    //une el nuevo nodo al frente del cache
     TempNode->data = strdup(data);
     TempNode->next = cache->head;
     cache->head = TempNode;
 
-    if (cache->size < cache->capacity)
+    if (cache->size < cache->capacity)//si el cache no esta lleno
     {
         cache->size++;
         printf("Dato '%s' agregado al cache. Tamano actual: %d\n", data, cache->size);
         update_data(data);
     }
-    else//aqui debe estar la funcion para reescribir el archivo data.txt
+    else//si el cache esta lleno
     {
         printf("Cache lleno. Reescribiendo data.txt...\n");
         Node_ *current = cache->head;
         Node_ *prev = NULL;
 
-        if(cache->head == NULL) {
+        if(cache->head == NULL)
+        {
             printf("Error: el cache esta vacio, no se puede eliminar ningun dato\n");
             return -1;
         }
         
-        while(current->next != NULL)
+        while(current->next != NULL)//recorre hasta el ultimo nodo
         {
             prev = current;
             current = current->next;
@@ -109,23 +118,23 @@ int lru_add(Cache_ *cache, char *data) // Comando para agregar un dato al cache
 
         printf("Eliminando dato menos reciente: '%s'\n", current->data);
 
-        if(prev != NULL)
+        if(prev != NULL)//desvincula el ultimo nodo
         {
             prev->next = NULL;
             free(current->data);
             free(current);
         }
 
-        rewrite_data(cache);
+        rewrite_data(cache);//reescribe el archivo data.txt
         printf("Cache lleno. Dato '%s' agregado al cache, el dato menos reciente fue eliminado.\n", data);
     }
-    update_cache(cache);
+    update_cache(cache);//actualiza el archivo metadata.txt
     return 0;
 }
 
 int lru_all(Cache_ *cache) // Comando para mostrar todos los datos en el cache
 {
-    if(cache == NULL || cache->head == NULL)
+    if(!cache || !cache->head)
     {
         printf("Cache vacio.\n");
         return -1;
@@ -134,19 +143,19 @@ int lru_all(Cache_ *cache) // Comando para mostrar todos los datos en el cache
     Node_ *current = cache->head;
     int index = 1;
 
-    while(current != NULL)
+    while(current != NULL)//recorre todos los nodos del cache
     {
-        if(current->data == NULL)
+        if(current->data == NULL)//si el dato es NULL
         {
             printf("[%d] Dato NULL\n", index);
         }
         else
         {
-            printf("%s", current->data);
+            printf("%s", current->data);//imprime el dato
         }
         if(index != cache->size)
         {
-            printf(" - ");
+            printf(" - ");//separa los datos con un guion
         }
         current = current->next;
         index++;
@@ -155,54 +164,43 @@ int lru_all(Cache_ *cache) // Comando para mostrar todos los datos en el cache
     return 0;
 }
 
-int lru_search(Cache_ *cache, char *data) // Comando de búsqueda de un dato en el cache
+int lru_search(Cache_ *cache, char *data) // Comando de busqueda de un dato en el cache
 {
-    if(cache == NULL || cache->head == NULL)
+   if(!cache || !cache->head)
     {
-        printf("Cache vacio. Resultado: -1\n");
+        printf("Cache vacio.\n");
         return -1;
     }
 
-    Node_ *current = cache->head;
-    int posicion = 1; // Empezamos a contar desde 1 para la posición
-
-    while(current != NULL)
+    if(search_data(cache, data) == 0)//si el dato no se encuentra
     {
-        // Compara el dato actual con el buscado
-        if(strcmp(current->data, data) == 0)
-        {
-            // ¡Encontrado! Imprime la posición 
-            printf("Dato '%s' encontrado en la posicion: %d\n", data, posicion);
-            return 0; // Termina la función con éxito
-        }
-        
-        // Si no es el dato, avanza al siguiente nodo
-        current = current->next;
-        posicion++; // Incrementa el contador de posición
+        printf("Dato '%s' no encontrado en el cache.\n", data);
+        return -1;
+    }
+    else//si el dato se encuentra
+    {
+        printf("Dato '%s' encontrado en el cache.\n", data);
     }
 
-    // Si el bucle termina, significa que no se encontró
-    printf("Dato '%s' no encontrado en el cache. Resultado: -1\n", data);
-    return -1; // Retorna -1 (como valor de error)
+    return -1;
 }
 
 int lru_get(Cache_ *cache, char *data) // Comandos para obtener un dato del cache
 {
-    if(cache == NULL || cache->head == NULL)
+    if(!cache || !cache->head)
     {
         printf("Cache vacio.\n");
         return -1;
     } 
 
-    if(search_data(cache, data) == 0)
+    if(search_data(cache, data) == 0)//si el dato no se encuentra
     {
         printf("Dato '%s' no encontrado en el cache.\n", data);
         return -1;
     }
-    //necesito un swap que recorra desde el nodo encontrado hasta el head
     
-    swap(cache, data);
-    rewrite_data(cache);
+    swap(cache, data);//mueve el dato al frente del cache
+    rewrite_data(cache);//reescribe el archivo data.txt
     printf("Dato '%s' movido a la posicion mas reciente del cache.\n", data);
 
     return 0;
@@ -210,9 +208,12 @@ int lru_get(Cache_ *cache, char *data) // Comandos para obtener un dato del cach
 
 int lru_exit() // Comando para salir y eliminar el cache
 {
-    printf("Eliminando cache y saliendo del programa...\n");
-
-    #ifdef _WIN32
+    if(!DirExists("cache"))//verifica si la carpeta cache existe
+    {
+        printf("No hay cache para eliminar.\n");
+        return -1;
+    }
+    #ifdef _WIN32//comando para eliminar la carpeta cache en Windows
         if(system("rmdir /S /Q cache") == -1)
         {
             printf("Error al eliminar la carpeta cache\n");
@@ -221,7 +222,7 @@ int lru_exit() // Comando para salir y eliminar el cache
         {
             printf("Carpeta cache eliminada con exito\n");
         }
-    #else
+    #else//comando para eliminar la carpeta cache en Linux/Mac
         if(system("rm -rf cache") == -1)
         {
             printf("Error al eliminar la carpeta cache\n");
